Rename the CRC struct to make the compiler happy
For some reason, the compiler was just not recognizing CRC as a valid identifier
in the Svc::CmdSequencerComponentImpl::FPrimeSequence::CRC struct definition.

Renaming it to something else random, CRCx, seemed to fix it.  I suppose
it must be a macro but I'm not sure where it's definied.

Coincidentally, it appears that CRCx just happens to already be defined
in another unrelated header that's not being used right now.